Add files via upload
Instructions for this install file: 

In order to use the model (Local Policing and Colorism, Uni-directional Influence.nlogo), you will need to install NetLogo.  The file associated with these comments is the installation software for NetLogo.  Download and install this file first.  Once you have installed NetLogo, download the model which has the file name of "Local Policing and Colorism, Uni-directional Influence.nlogo".   Once you have downloaded the model, click on the icon for the model and follow the instructions for using the model.  The instructions can be found under the "Info" tab.